Parser_Deployment/requirements.txt
llama-parse
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3